% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoherenceChIA.R
\name{active.gene.coherence}
\alias{active.gene.coherence}
\title{Given a ChIA object, determine which genes are active and which are not.}
\usage{
active.gene.coherence(chia.obj)
}
\arguments{
\item{chia.obj}{The base ChIA object on which to assess gene activity.}
}
\value{
A 2-level vector with values "Inactive", "Active" or NA for each node.
}
\description{
Given a ChIA object, determine which genes are active and which are not.
}

