% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateChIA.R
\name{add.gene.annotation}
\alias{add.gene.annotation}
\title{Adds gene-specific annotations to chia.obj.}
\usage{
add.gene.annotation(chia.obj, annotation.df, label)
}
\arguments{
\item{chia.obj}{A chia object to annotate.}

\item{annotation.df}{A data frame where the first column is a gene symbol, and the
second one is a value to be associated to that gene.}

\item{label}{The label the annotation should be given in the chia object.}
}
\value{
The ChIA object with the added annotation.
}
\description{
Given a data frame where the first column are gene symbols, and the second column
are values associated with the given gene, this function finds teh gene representatives
nodes for all symbols and attaches the given annotation to it.
}

