% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChIAParameters.R
\name{add_encode_data}
\alias{add_encode_data}
\title{Retrieve whatever data it can from ENCODE and add it to the ChIA parameter object.}
\usage{
add_encode_data(chia.params)
}
\arguments{
\item{chia.params}{The ChIA parameters object for which additional annotations must be fetched from ENCODE.}
}
\value{
An environment that can be passed to process_chia_pet, annotate_chia.pet or analyze_chia_pet.
}
\description{
Retrieve whatever data it can from ENCODE and add it to the ChIA parameter object.
}

