% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisStepsChIA.R
\name{analyze.chromatin.states}
\alias{analyze.chromatin.states}
\title{Analyze the chromatin state of ChIA-PET data}
\usage{
analyze.chromatin.states(chia.obj, chia.params = NULL,
  output.dir = "output")
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned
by \code{\link{annotate.chia}}}

\item{chia.params}{Original chia.params used to build chia.obj, used to 
determine genomic backgrounds.}

\item{output.dir}{The name of the directory where to save the graphs.}
}
\description{
Analyzes ChIA-PET data and produces graphs according to its chromatin states.
}

