% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisStepsChIA.R
\name{analyze_central_nodes}
\alias{analyze_central_nodes}
\title{Performs all possible analysis regarding the central nodes.}
\usage{
analyze_central_nodes(chia.obj, output.dir = ".")
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate_chia}}.}

\item{output.dir}{The directory where output should be saved.}
}
\description{
Performs all possible analysis regarding the central nodes.
}

