% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisStepsChIA.R
\name{analyze_components}
\alias{analyze_components}
\title{Analyze the components of ChIA-PET data}
\usage{
analyze_components(chia.obj, output.dir = "output")
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate_chia}}.}

\item{output.dir}{The name of the directory where to save the graphs.}
}
\description{
Analyzes ChIA-PET data and produces graphs according to the topology of the nodes in the whole set
of regions and in the separated components: \describe{
\item{Log2(size of component) vs Log2(Number of components).pdf}{A plot of the number_of_components according to the size of the components.}
\item{Component table.txt}{A file with the information about TSS for each component.}
\item{Proportion of TSS in low connectivity nodes.pdf}{A plot of the proportion of TSS in nodes with low connectivity (component with 5 nodes or less).}
\item{Proportion of TSS in high connectivity nodes.pdf}{A plot of the proportion of TSS in nodes with high connectivity (component with more than 5 nodes).}}
}

