% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalysisStepsChIA.R
\name{analyze_misc}
\alias{analyze_misc}
\title{Performs analyses which do not fit in any of the other sub-analyses.}
\usage{
analyze_misc(chia.obj, output.dir = ".")
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate_chia}}.}

\item{output.dir}{The directory where output should be saved.}
}
\description{
Performs analyses which do not fit in any of the other sub-analyses.
}

