% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateChIA.R
\name{annotate.chia}
\alias{annotate.chia}
\title{Annotate "chia.obj", given as parameter.}
\usage{
annotate.chia(chia.obj, chia.param, output.dir = ".", verbose = TRUE,
  skip.centrality = FALSE)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}

\item{chia.param}{An object returned by build.chia.params indicating which data should be used for annotating the ChIA object.}

\item{output.dir}{The name of the directory where to write the selected annotations.}

\item{verbose}{If TRUE, console output is suppressed.}

\item{skip.centrality}{If TRUE, centrality scores are not computed.}
}
\value{
The annotated "\code{chia.obj}".
}
\description{
Annotate "chia.obj", given as parameter.
}

