% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{apply_by_component}
\alias{apply_by_component}
\title{Apply a metric function to each component in the ChIA object.}
\usage{
apply_by_component(chia.obj, metric.function, ...)
}
\arguments{
\item{chia.obj}{The chia object whose compoennts should have their metrics evaluated.}

\item{metric.function}{The metric function to apply to all components.}
}
\value{
A list containing the measures metrics for all components.
}
\description{
Apply a metric function to each component in the ChIA object.
}

