% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{apply_single_metric_by_component}
\alias{apply_single_metric_by_component}
\title{Apply a metric function returning a single metric to all components of a ChIA object,
and return teh resulting values as a single multi-value metric.}
\usage{
apply_single_metric_by_component(metric.function, metric.name)
}
\arguments{
\item{metric.function}{The metric function to apply to all components.}

\item{metric.name}{The name for the resulting combined metric.}
}
\value{
A list containing the measures metrics for all components.
}
\description{
Apply a metric function returning a single metric to all components of a ChIA object,
and return teh resulting values as a single multi-value metric.
}

