% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateChIA.R
\name{associate_regions_by_name}
\alias{associate_regions_by_name}
\title{Associate named regions to a ChIA-PET object.}
\usage{
associate_regions_by_name(chia.obj, annotation.regions, annotation.name)
}
\arguments{
\item{annotation.regions}{A \linkS4class{GRanges} object with the named 
regions to be used for annotation.}

\item{annotation.name}{The name to be given to the new column within the 
ChIA-PET object's annotations.}

\item{The}{ChIA-PET object to annotate.}
}
\value{
The ChIA-PET object, with added annotations.
}
\description{
Associate named regions to a ChIA-PET object.
}

