% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{average_component_size}
\alias{average_component_size}
\title{Return the mean component size of a CHIA object.}
\usage{
average_component_size(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
The mean component size of the chia object.
}
\description{
Return the mean component size of a CHIA object.
}

