% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricsChIA.R
\name{boolean_count}
\alias{boolean_count}
\title{Finds which transcription factors are present}
\usage{
boolean_count(chia.obj, variable.name, proportion = FALSE)
}
\arguments{
\item{variable.name}{The name of the boolean variable whose proportion should be calculated.}

\item{chia.subset}{A list containing a graph and ChIA-PET regions, as returned by \code{\link{chia_vertex_subset}}.}
}
\value{
The count/proportion of nodes whose attribute named "variable.name" is TRUE.
}
\description{
Finds which transcription factors are present
}

