% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricsChIA.R
\name{calculate.tf.presence}
\alias{calculate.tf.presence}
\title{Calculates which number/proportion of nodes harbor each transcription factor.}
\usage{
calculate.tf.presence(chia.obj, proportion = TRUE)
}
\arguments{
\item{chia.obj}{The chia.obj whose network must be analyzed.}

\item{proportion}{Should the proportion be returned instead of the number of regions?}
}
\value{
A number/proportion of regions occupied by each transcription factor.
}
\description{
Calculates which number/proportion of nodes harbor each transcription factor.
}

