% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategorizeChIA.R
\name{categorize.by.centrality}
\alias{categorize.by.centrality}
\title{Creates categories based on the centrality of the nodes}
\usage{
categorize.by.centrality(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate.chia}}.}
}
\value{
A data frame in which each column corresponds to a category. The data frame is as long as the
"Regions" element of chia.obj, and the values indicate wheter or not a region belongs to a category.
}
\description{
Creates categories based on the centrality of the nodes
}

