% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategorizeChIA.R
\name{categorize_by_component}
\alias{categorize_by_component}
\title{Creates categories based on the component of each node.}
\usage{
categorize_by_component(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate_chia}}.}
}
\value{
A data frame in which each column corresponds to a category. The data frame is as long as the
"Regions" element of chia.obj, and the values indicate wheter or not a region belongs to a category.
}
\description{
Creates categories based on the component of each node.
}

