% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategorizeChIA.R
\name{categorize_by_tf}
\alias{categorize_by_tf}
\title{Creates categories based on the presence of certain transcription factors.}
\usage{
categorize_by_tf(chia.obj, tf.names = NULL, add.all = FALSE)
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate_chia}}.}

\item{tf.names}{The transcription factors to be used for the categories. If NULL, all available TFs are used.}

\item{add.all}{IF TRUE, a category containing all nodes is added.}
}
\value{
A list containing a set of indices, one for each apssed TF combination, indicating which regions
  bind the given TFs.
}
\description{
Creates categories based on the presence of certain transcription factors.
}

