% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentChIA.R
\name{chia.plot.enrichment}
\alias{chia.plot.enrichment}
\title{Given a data-frame combining the results of multiple calls to node.enrichment, plot those enrichments.}
\usage{
chia.plot.enrichment(enrichment.df, filename, label = "Category")
}
\arguments{
\item{enrichment.df}{A dataframe combining the results of multiple node.enrichment calls.}

\item{filename}{The name of the file where the plot should be saved.}

\item{label}{A label for the categories represented by enrichment.df's rows.}
}
\description{
Given a data-frame combining the results of multiple calls to node.enrichment, plot those enrichments.
}

