% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{chia_contract}
\alias{chia_contract}
\title{Contracts an interaction network, keeping only the vertices in indices.}
\usage{
chia_contract(chia.obj, indices)
}
\arguments{
\item{chia.obj}{The ChIA object to be contracted.}

\item{chip.names}{The indices of the regions to keep in the output chia object.}
}
\value{
A contracted ChIA object.
}
\description{
Suppose we have a ring network, 1->2->3->4->5->1.
Now suppose the passed indices are 2, 4 and 5.
Thus, nodes 1 and 3 would be removed, and new edges from 2 to 4 and 5 to
2 would be added. The new network would be:
2->4->5->2.
}

