% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricsChIA.R
\name{chia_plot_metrics}
\alias{chia_plot_metrics}
\title{Plots a subset of ChIA-PET nodes}
\usage{
chia_plot_metrics(chia.obj, metric.function, node.categories, x.lab = NULL,
  y.lab = NULL, title = NULL, graph.type = "line", facet.rows = NULL,
  facet.cols = NULL, file.out = NULL, ...)
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate_chia}}.}

\item{metric.function}{A function taking a chia.obj parameter and returning metrics in order to construct the plot.}

\item{node.categories}{Indices of nodes representing subsets of these nodes.}

\item{x.lab}{The label to add to the x axis of the gaph.}

\item{y.lab}{The label to add to the y axis of the graph.}

\item{title}{The title to add to the graph.}

\item{graph.type}{Either "line", "histogram", "heatmap".}

\item{facet.rows}{Optionnal graphic parameter.}

\item{facet.cols}{Optionnal graphic parameter.}

\item{file.out}{The name of the file where to save the plot, or NULL if none should be saved.}

\item{\dots}{Paramters to pass to metric.function.}
}
\value{
The plot.
}
\description{
Plots a subset of ChIA-PET nodes
}

