% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{chia_vertex_subset}
\alias{chia_vertex_subset}
\title{Subset a CHIA object.}
\usage{
chia_vertex_subset(chia.obj, indices)
}
\arguments{
\item{chia.obj}{The chia object to be subset.}

\item{indices}{The indices of the vertices to be kept.}
}
\value{
A chia object containing only the selected vertices.
}
\description{
Subset a CHIA object.
}

