% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoherenceChIA.R
\name{coherence_test}
\alias{coherence_test}
\title{Determines if certain attributes are coherent within certain categories of nodes.}
\usage{
coherence_test(chia.obj, coherence.function, node.categories, m.hyper,
  output.file = NULL, ...)
}
\arguments{
\item{chia.obj}{The base ChIA object on which to assess intra-category coherence.}

\item{coherence.function}{A function which takes in a ChIA object and returns a 
vector of a 2-level factor of the same length as the number fo regions in the 
ChIA object.}

\item{node.categories}{A data-frame with a number of rows equal to the number of regions
in chia.obj, where each column represents a set of indices indicating which nodes belong
in a given category.}

\item{m.hyper}{The m parameter for the hypergeometric test, namely the total number of coherent genes.}

\item{output.file}{The name of the file where the resulting graph should be saved.}
}
\value{
A list containing the computed data and the generated plot object.
}
\description{
Determines if certain attributes are coherent within certain categories of nodes.
}

