% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricsChIA.R
\name{count_cut}
\alias{count_cut}
\title{Counts the occurences of categories made with a variable}
\usage{
count_cut(chia.subset, variable.name, proportion = TRUE, ...)
}
\arguments{
\item{chia.subset}{A list containing a graph and ChIA-PET regions, as returned by \code{\link{chia_vertex_subset}}.}

\item{variable.name}{The name of the column containing a factor.}

\item{proportion}{Should the number of occurences be converted to proportions?}
}
\value{
A named vector (table) with the number of occurences (or proportions) of each category.
}
\description{
Counts the occurences of categories made with a variable
}

