% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoherenceChIA.R
\name{expression_coherence}
\alias{expression_coherence}
\title{Categorizes regions according to their expression level.}
\usage{
expression_coherence(chia.obj, top, bottom)
}
\arguments{
\item{chia.obj}{The chia object whose regions must be classified.}

\item{top}{The threshold above which a region is classified as "Top".}

\item{bottom}{The threshold under which a region is classified as "Bottom".}
}
\value{
A vector with a 2-level factor, where regions are classified either
  as "Top", "Bottom" or NA according to their expression level.
}
\description{
Given expression threshold, this function classifies all regions as 
belonging to a gene "below" the bottom expression level or "above" the top
expression level.
}

