% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeChIA.R
\name{fancy.topology}
\alias{fancy.topology}
\title{Represents contact-frequencies between different categories of elements
as a stylized graph.}
\usage{
fancy.topology(chia.obj, col.name,
  file.name = paste0("Abstract inter-category contact graph for ", col.name,
  ".pdf"))
}
\arguments{
\item{chia.obj}{A list containing the annotated ChIA-PET data, as returned by \code{\link{annotate.chia}}.}

\item{col.name}{The name of the column on which the contact graph should be based.}

\item{file.name}{The name of the output file.}
}
\description{
Creates a graph where each node represents a node category from the original chia.obj,
and edge width represents the frequency of those contacts. Node size is
also proportional to the number of nodes in the given category.
}

