% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoherenceChIA.R
\name{fold.change.coherence}
\alias{fold.change.coherence}
\title{Generates a function which takes a chia object, and determine which of its nodes
have an positive or negative fold-change based on the given column name.}
\usage{
fold.change.coherence(fc.column)
}
\arguments{
\item{fc.column}{The column the new function should check for fold-change values.}
}
\value{
A function taking a ChIA object, and returning a 2-level vector with 
  "Upregulated", "Downregulated" or NA depending on the value of the given 
  fold-change column.
}
\description{
Generates a function which takes a chia object, and determine which of its nodes
have an positive or negative fold-change based on the given column name.
}

