% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricsChIA.R
\name{functor.constructor}
\alias{functor.constructor}
\title{Builds a function object for extracting metrics using a variable name.}
\usage{
functor.constructor(method, variable.name, proportion = FALSE)
}
\arguments{
\item{method}{The method to be used by the resulting function object. Valid values could be
levels.count, count.cut, boolean.count, etc.}

\item{variable.name}{The name of the variable whose proportion/counts should be calculated.}

\item{proportion}{If true, proportions are returned instead of counts.}
}
\value{
A function to calculate the desired metric.
}
\description{
Builds a function object for extracting metrics using a variable name.
}

