% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{genes_by_component}
\alias{genes_by_component}
\title{Return the number genes per component in the CHIA object.}
\usage{
genes_by_component(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
The number of active genes represented in the chia object.
}
\description{
Return the number genes per component in the CHIA object.
}

