% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{get.annotated.edges}
\alias{get.annotated.edges}
\title{Return a data-frame representing all edges and the annotation of the 
vertices it connects.}
\usage{
get.annotated.edges(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by 
\code{\link{load.chia}}.}
}
\value{
A data-frame representing all edges of the graph and their annotation.
}
\description{
Return a data-frame representing all edges and the annotation of the 
vertices it connects.
}

