% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{get.chips.names}
\alias{get.chips.names}
\title{Obtain the names of ChIP data sets for which annotation is available.}
\usage{
get.chips.names(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
A vector containing the names of the ChIP data sets for which annotation is available.
}
\description{
Obtain the names of ChIP data sets for which annotation is available.
}

