% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{get.granges}
\alias{get.granges}
\title{Return the right part of the ChIA-PET data.}
\usage{
get.granges(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
A \linkS4class{GRanges} object with the \code{Regions} from the "\code{chia.obj}" parameter corresponding the the "right side".
of the original data.
}
\description{
Return the right part of the ChIA-PET data.
}

