% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{get.histones.names}
\alias{get.histones.names}
\title{Obtain the names of histone marks for which annotation is available.}
\usage{
get.histones.names(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
A vector containing the names of the histone marks for which annotation is available.
}
\description{
Obtain the names of histone marks for which annotation is available.
}

