% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{get.histones.percent}
\alias{get.histones.percent}
\title{Obtain the matrix of histone marks binding as a percentage of the ChIA regions..}
\usage{
get.histones.percent(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
A matrix indicating which histone marks binds to which regions.
}
\description{
Obtain the matrix of histone marks binding as a percentage of the ChIA regions..
}

