% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{get.histones}
\alias{get.histones}
\title{Obtain the matrix of histone marks binding.}
\usage{
get.histones(chia.obj, overlap.threshold = 0)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
A matrix indicating which histone marks binds to which regions.
}
\description{
Obtain the matrix of histone marks binding.
}

