% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{get.polymerases.percent}
\alias{get.polymerases.percent}
\title{Obtain the matrix of polymerase binding as a percentage of the ChIA regions.}
\usage{
get.polymerases.percent(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
A matrix indicating which polymerases binds to which regions.
}
\description{
Obtain the matrix of polymerase binding as a percentage of the ChIA regions.
}

