% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{get_all_chip_proportion}
\alias{get_all_chip_proportion}
\title{Obtain the proportion of regions intersecting a ChIP dataset, for all ChIP datasets.}
\usage{
get_all_chip_proportion(chia.obj)
}
\arguments{
\item{chia.obj}{A ChIA-PET object.}
}
\value{
A named numeric vector containing the proportion of regions intersecting the ChIP datasets
}
\description{
Obtain the proportion of regions intersecting a ChIP dataset, for all ChIP datasets.
}

