% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{get_all_statistics}
\alias{get_all_statistics}
\title{Returns a list of all statistics for a given ChIA object.}
\usage{
get_all_statistics(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
A named list of calculated properties for the given ChIA object.
}
\description{
Returns a list of all statistics for a given ChIA object.
}

