% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{get_chips_names}
\alias{get_chips_names}
\title{Obtain the names of ChIP data sets for which annotation is available.}
\usage{
get_chips_names(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
A vector containing the names of the ChIP data sets for which annotation is available.
}
\description{
Obtain the names of ChIP data sets for which annotation is available.
}

