% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{get_gene_annotation}
\alias{get_gene_annotation}
\title{Returns the region annotations for the representative regions for the given genes.}
\usage{
get_gene_annotation(chia.obj, genenames)
}
\arguments{
\item{chia.obj}{The ChIA object whose regions must be assessed.}

\item{genenames}{The genes whose representative annotation must be returned.}
}
\value{
A data-frame containign the annotation for the selected genes.
}
\description{
Returns the region annotations for the representative regions for the given genes.
}

