% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{get_polymerases}
\alias{get_polymerases}
\title{Obtain the matrix of polymerase binding.}
\usage{
get_polymerases(chia.obj, overlap.threshold = 0)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
A matrix indicating which polymerases binds to which regions.
}
\description{
Obtain the matrix of polymerase binding.
}

