% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{get_tf}
\alias{get_tf}
\title{Obtain the matrix of TF binding.}
\usage{
get_tf(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
A matrix indicating which TF binds to which regions.
}
\description{
Obtain the matrix of TF binding.
}

