% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{get_tf_names}
\alias{get_tf_names}
\title{Obtain the name of transcription factors for which annotation is available.}
\usage{
get_tf_names(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
A vector containing the names of the transcription factors for which annotation is available.
}
\description{
Obtain the name of transcription factors for which annotation is available.
}

