% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{has.degree}
\alias{has.degree}
\title{Determines if the given chia.obj has associated node degrees.}
\usage{
has.degree(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
True if the object has associated node degrees.
}
\description{
Determines if the given chia.obj has associated node degrees.
}

