% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{has.expression.levels}
\alias{has.expression.levels}
\title{Determines if the given chia.obj has associated expression levels.}
\usage{
has.expression.levels(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
True if the object has associated expression levels.
}
\description{
Determines if the given chia.obj has associated expression levels.
}

