% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{has.fitness}
\alias{has.fitness}
\title{Determines if the given chia.obj has TF binding information.}
\usage{
has.fitness(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
True if the object has TF binding information.
}
\description{
Determines if the given chia.obj has TF binding information.
}

