% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{has.gene.representative}
\alias{has.gene.representative}
\title{Determines if the given chia.obj has associated gene representatives.}
\usage{
has.gene.representative(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
True if the object has associated gene representatives.
}
\description{
Determines if the given chia.obj has associated gene representatives.
}

