% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{has.histone.marks}
\alias{has.histone.marks}
\title{Determines if the given chia.obj has histone marks binding information.}
\usage{
has.histone.marks(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
True if the object has histone marks binding information.
}
\description{
Determines if the given chia.obj has histone marks binding information.
}

