% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{has.polymerases}
\alias{has.polymerases}
\title{Determines if the given chia.obj has polymerase binding information.}
\usage{
has.polymerases(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
True if the object has polymerase binding information.
}
\description{
Determines if the given chia.obj has polymerase binding information.
}

