% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{has_centrality}
\alias{has_centrality}
\title{Determines if the given chia.obj has associated node centralities.}
\usage{
has_centrality(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
True if the object has associated node centralities.
}
\description{
Determines if the given chia.obj has associated node centralities.
}

