% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{has_chrom_state}
\alias{has_chrom_state}
\title{Determines if the given chia.obj has associated chromatin states.}
\usage{
has_chrom_state(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
True if the object has associated chromatin states.
}
\description{
Determines if the given chia.obj has associated chromatin states.
}

