% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{has_components}
\alias{has_components}
\title{Determines if the given chia.obj has associated components.}
\usage{
has_components(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
True if the object has associated components.
}
\description{
Determines if the given chia.obj has associated components.
}

