% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{has_expression_levels}
\alias{has_expression_levels}
\title{Determines if the given chia.obj has associated expression levels.}
\usage{
has_expression_levels(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
True if the object has associated expression levels.
}
\description{
Determines if the given chia.obj has associated expression levels.
}

