% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{has_gene_representative}
\alias{has_gene_representative}
\title{Determines if the given chia.obj has associated gene representatives.}
\usage{
has_gene_representative(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
True if the object has associated gene representatives.
}
\description{
Determines if the given chia.obj has associated gene representatives.
}

