% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{has_histone_marks}
\alias{has_histone_marks}
\title{Determines if the given chia.obj has_histone_marks binding information.}
\usage{
has_histone_marks(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
True if the object has_histone_marks binding information.
}
\description{
Determines if the given chia.obj has_histone_marks binding information.
}

