% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{has_polymerases}
\alias{has_polymerases}
\title{Determines if the given chia.obj has polymerase binding information.}
\usage{
has_polymerases(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
True if the object has polymerase binding information.
}
\description{
Determines if the given chia.obj has polymerase binding information.
}

