% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeChIA.R
\name{histogram_essential_genes}
\alias{histogram_essential_genes}
\title{Histogram of the proportion of "essential genes", by connectivity categories.}
\usage{
histogram_essential_genes(chia.obj, essential.genes, label.x, output.dir)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{annotate_chia}}.}

\item{essential.genes}{The ID's of the genes to anlayze. The ID's refer the the ChIA-PET ID's of the nodes.}

\item{label.x}{The title of the histogram.}

\item{output.dir}{The name of the directory where to save the plot.}
}
\description{
Histogram of the proportion of "essential genes", by connectivity categories.
}

