% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{is_gene_representative}
\alias{is_gene_representative}
\title{Returns a logical vector indicating which nodes are gene representatives.}
\usage{
is_gene_representative(chia.obj)
}
\arguments{
\item{chia.obj}{The ChIA object whose regions must be assessed.}
}
\value{
A logical vector indicating which nodes are gene representatives.
}
\description{
Returns a logical vector indicating which nodes are gene representatives.
}

