% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricsChIA.R
\name{level_counts}
\alias{level_counts}
\title{Counts the occurences of each factor}
\usage{
level_counts(chia.subset, variable.name, proportion = TRUE)
}
\arguments{
\item{chia.subset}{A list containing a graph and ChIA-PET regions, as returned by \code{\link{chia_vertex_subset}}.}

\item{variable.name}{The name of the column containing a factor.}

\item{proportion}{Should the number of occurences be converted to proportions?}
}
\value{
A named vector (table) with the number of occurences (or proportions) of each factor in the data.
}
\description{
Counts the occurences of each factor
}

