% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricsChIA.R
\name{metric.list.to.df}
\alias{metric.list.to.df}
\title{Converts a list of metrics into a data-frame for plotting.}
\usage{
metric.list.to.df(metric.list)
}
\arguments{
\item{The}{list of metrics returned by mulitple call to a metric.function.}
}
\value{
A data-frame representing the list values in a structure appropriate for ggplot.
}
\description{
Converts a list of metrics into a data-frame for plotting.
}

