% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentChIA.R
\name{node_enrichment}
\alias{node_enrichment}
\title{Given two sets of nodes, determine if one shows enrichment.}
\usage{
node_enrichment(chia.obj, hit.functor, draw.functor)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}

\item{hit.functor}{A functor selecting the "white" elements.}

\item{draw.functor}{A functor selecting the "drawn" elements.}
}
\value{
A vector with enrichment properties.
}
\description{
Given two sets of nodes, determine if one shows enrichment.
}

