% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{nodes.with.polymerase}
\alias{nodes.with.polymerase}
\title{Obtain the indices of nodes bearing any or all of a set of transcription factors.}
\usage{
nodes.with.polymerase(chia.obj, polymerase.names,
  how.many = length(polymerase.names))
}
\arguments{
\item{chia.obj}{A ChIA-PET object.}

\item{how.many}{How many of the selected TFs should be present to return a hit? By default,
all TFs must be present.}

\item{tf.names}{The names of the transcription factor to look for.}
}
\value{
A logical vector indicating which regions bind the given transcription factor.
}
\description{
Obtain the indices of nodes bearing any or all of a set of transcription factors.
}

