% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{nodes_with_tf}
\alias{nodes_with_tf}
\title{Obtain the indices of nodes bearing any or all of a set of transcription factors.}
\usage{
nodes_with_tf(chia.obj, tf.names, how.many = length(tf.names))
}
\arguments{
\item{chia.obj}{A ChIA-PET object.}

\item{tf.names}{The names of the transcription factor to look for.}

\item{how.many}{How many of the selected TFs should be present to return a hit? By default,
all TFs must be present.}
}
\value{
A logical vector indicating which regions bind the given transcription factor.
}
\description{
Obtain the indices of nodes bearing any or all of a set of transcription factors.
}

