% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{number.of.polymerases}
\alias{number.of.polymerases}
\title{Obtain the number of polymerases from a given list that bind all regions within a ChIA-PET object.}
\usage{
number.of.polymerases(chia.obj, polymerase.names)
}
\arguments{
\item{chia.obj}{A ChIA-PET object.}

\item{polymerase.names}{The names of the transcription factor to look for.}
}
\value{
An integer vector containing the number of polymerases in the list binding each region.
}
\description{
Obtain the number of polymerases from a given list that bind all regions within a ChIA-PET object.
}

