% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{number.of.tfs}
\alias{number.of.tfs}
\title{Obtain the number of TFs from a given list that bind all regions within a ChIA-PET object.}
\usage{
number.of.tfs(chia.obj, tf.names)
}
\arguments{
\item{chia.obj}{A ChIA-PET object.}

\item{tf.names}{The names of the transcription factor to look for.}
}
\value{
An integer vector containing the number of TF in the list binding each region.
}
\description{
Obtain the number of TFs from a given list that bind all regions within a ChIA-PET object.
}

