% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{number_active_genes}
\alias{number_active_genes}
\title{Return the number of active genes represented in a CHIA object.}
\usage{
number_active_genes(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
The number of active genes represented in the chia object.
}
\description{
Return the number of active genes represented in a CHIA object.
}

