% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{number_of_components}
\alias{number_of_components}
\title{Return the number_of_components in a CHIA object.}
\usage{
number_of_components(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
The number_of_components in the chia object.
}
\description{
Return the number_of_components in a CHIA object.
}

