% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{number_of_contacts}
\alias{number_of_contacts}
\title{Return the number_of_contacts in a CHIA object.}
\usage{
number_of_contacts(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
The number_of_contacts in the chia object.
}
\description{
Return the number_of_contacts in a CHIA object.
}

