% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{number_of_polymerases}
\alias{number_of_polymerases}
\title{Obtain the number_of_polymerases from a given list that bind all regions within a ChIA-PET object.}
\usage{
number_of_polymerases(chia.obj, polymerase.names)
}
\arguments{
\item{chia.obj}{A ChIA-PET object.}

\item{polymerase.names}{The names of the transcription factor to look for.}
}
\value{
An integer vector containing the number_of_polymerases in the list binding each region.
}
\description{
Obtain the number_of_polymerases from a given list that bind all regions within a ChIA-PET object.
}

