% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{process_chia_pet}
\alias{process_chia_pet}
\title{Analyze ChIA-PET data and produce graphs.}
\usage{
process_chia_pet(input.chia, chia.param, output.dir = "output",
  verbose = TRUE)
}
\arguments{
\item{input.chia}{The file containing processed ChIA-PET data.}

\item{output.dir}{The name of the directory where output should be saved.}
}
\value{
The annotated chia.obj.
}
\description{
Analyze ChIA-PET data and produce graphs.
}

