% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{proportion_active_genes}
\alias{proportion_active_genes}
\title{Return the proportion of active genes among all genes in a CHIA object.}
\usage{
proportion_active_genes(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
The proportion of active genes among all genes in the chia object.
}
\description{
Return the proportion of active genes among all genes in a CHIA object.
}

