% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{proportion_genes}
\alias{proportion_genes}
\title{Return the proportion of nodes representing genes in a CHIA object.}
\usage{
proportion_genes(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
The number of active genes represented in the chia object.
}
\description{
Return the proportion of nodes representing genes in a CHIA object.
}

