% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricsChIA.R
\name{region.column.functor}
\alias{region.column.functor}
\title{Builds a function object for extracting a single named annotation from the ChIA object.}
\usage{
region.column.functor(var.name)
}
\arguments{
\item{var.name}{The name of the Regions column to be returned.
levels.count, count.cut, boolean.count, etc.}
}
\value{
A function which extracts the given Regions column as a list.
}
\description{
Builds a function object for extracting a single named annotation from the ChIA object.
}

