% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricsChIA.R
\name{region_column_functor}
\alias{region_column_functor}
\title{Builds a function object for extracting a single named annotation from the ChIA object.}
\usage{
region_column_functor(var.name)
}
\arguments{
\item{var.name}{The name of the Regions column to be returned.
levels.count, count_cut, boolean_count, etc.}
}
\value{
A function which extracts the given Regions column as a list.
}
\description{
Builds a function object for extracting a single named annotation from the ChIA object.
}

