% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{regions.in.gene.network}
\alias{regions.in.gene.network}
\title{Determiens which regions of a ChIA object are part of gene networks.}
\usage{
regions.in.gene.network(chia.obj, min.gene = 2)
}
\arguments{
\item{chia.obj}{The ChIA object whose regions must be assessed.}

\item{min.gene}{The minimum number of gene in a component for it to be considered
a gene network.}
}
\value{
A logical vector indicating whcih regiosn are part of a gene network.
}
\description{
Determiens which regions of a ChIA object are part of gene networks.
}

