% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{select.by.component.functor}
\alias{select.by.component.functor}
\title{Generate a function similar to select.by.components where the components.ids parameter is bound.}
\usage{
select.by.component.functor(component.ids)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
A function which accepts a ChIA object and returns the subset of nodes whose
  component is in component.ids.
}
\description{
Generate a function similar to select.by.components where the components.ids parameter is bound.
}

