% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{select.by.components}
\alias{select.by.components}
\title{Select all nodes which are in the given components.}
\usage{
select.by.components(chia.obj, component.ids)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
A subset of chia.obj containing only nodes which are in the given components.
}
\description{
Select all nodes which are in the given components.
}

