% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{select.central.nodes}
\alias{select.central.nodes}
\title{Select all nodes which are central.}
\usage{
select.central.nodes(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
A subset of chia.obj containing only nodes which are central.
}
\description{
Select all nodes which are central.
}

