% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{select.factories}
\alias{select.factories}
\title{Select all nodes which are in a factory.}
\usage{
select.factories(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
A subset of chia.obj containing only nodes which are in a factory.
}
\description{
Select all nodes which are in a factory.
}

