% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{select.gene.networks}
\alias{select.gene.networks}
\title{Returns a subset of a ChIA object containing only components classified as gene networks.}
\usage{
select.gene.networks(chia.obj, min.gene = 2)
}
\arguments{
\item{chia.obj}{The ChIA object whose regions must be assessed.}

\item{min.gene}{The minimum number of gene in a component for it to be considered
a gene network.}
}
\value{
A subsetted chia object containing only the gene networks.
}
\description{
Returns a subset of a ChIA object containing only components classified as gene networks.
}

