% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{select.gene.representative}
\alias{select.gene.representative}
\title{Select all nodes which are gene representatives.}
\usage{
select.gene.representative(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
A subset of chia.obj containing only nodes which are gene representatives.
}
\description{
Select all nodes which are gene representatives.
}

