% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{select_by_component_functor}
\alias{select_by_component_functor}
\title{Generate a function similar to select_by_components where the components.ids parameter is bound.}
\usage{
select_by_component_functor(component.ids)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
A function which accepts a ChIA object and returns the subset of nodes whose
  component is in component.ids.
}
\description{
Generate a function similar to select_by_components where the components.ids parameter is bound.
}

