% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{select_by_components}
\alias{select_by_components}
\title{Select all nodes which are in the given components.}
\usage{
select_by_components(chia.obj, component.ids)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
A subset of chia.obj containing only nodes which are in the given components.
}
\description{
Select all nodes which are in the given components.
}

