% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{select_by_histone}
\alias{select_by_histone}
\title{Returns a subset of a ChIA object containing only nodes with the specified 
transcription factors.}
\usage{
select_by_histone(chia.obj, histone.names)
}
\arguments{
\item{chia.obj}{The ChIA object whose regions must be assessed.}

\item{histone.names}{The histone marks whose regions must be returned.}
}
\value{
A subset of the ChIA object containing only nodes with the specified 
  histone marks.
}
\description{
Returns a subset of a ChIA object containing only nodes with the specified 
transcription factors.
}

