% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{select_by_tf}
\alias{select_by_tf}
\title{Returns a subset of a ChIA object containing only nodes with the specified 
transcription factors.}
\usage{
select_by_tf(chia.obj, tf.names)
}
\arguments{
\item{chia.obj}{The ChIA object whose regions must be assessed.}

\item{tf.names}{The transcription factors whose regions must be returned.}
}
\value{
A subset of the ChIA object containing only nodes with the specified 
  transcription factors.
}
\description{
Returns a subset of a ChIA object containing only nodes with the specified 
transcription factors.
}

