% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{select_central_nodes}
\alias{select_central_nodes}
\title{Select all nodes which are central.}
\usage{
select_central_nodes(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
A subset of chia.obj containing only nodes which are central.
}
\description{
Select all nodes which are central.
}

