% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{select_factories}
\alias{select_factories}
\title{Select all nodes which are in a factory.}
\usage{
select_factories(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
A subset of chia.obj containing only nodes which are in a factory.
}
\description{
Select all nodes which are in a factory.
}

