% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{select_from_chia_functor}
\alias{select_from_chia_functor}
\title{Generate a function which selects all nodes whose ID is present in the passed-in chia.obj.}
\usage{
select_from_chia_functor(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
A function which accepts a ChIA object and returns the subset of nodes whose
  ID are in the ChIA object passed to this function.
}
\description{
Generate a function which selects all nodes whose ID is present in the passed-in chia.obj.
}

