% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationsChIA.R
\name{select_gene_representative}
\alias{select_gene_representative}
\title{Select all nodes which are gene representatives.}
\usage{
select_gene_representative(chia.obj)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load_chia}}.}
}
\value{
A subset of chia.obj containing only nodes which are gene representatives.
}
\description{
Select all nodes which are gene representatives.
}

