% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{shortest_distance}
\alias{shortest_distance}
\title{Returns the shortest_distances between a set of nodes and another one.}
\usage{
shortest_distance(chia.obj, from, to)
}
\arguments{
\item{chia.obj}{The ChIA object on which distances must be computed.}

\item{from}{A vector of indices describing the starting nodes in the distance calculations.}

\item{to}{A vector of indices describing the destination nodes in the distance calculations.}
}
\value{
A vector containing the distance from all "from" nodes to all "to" nodes.
}
\description{
Returns the shortest_distances between a set of nodes and another one.
}

