% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricsChIA.R
\name{subset.counts}
\alias{subset.counts}
\title{Counts the number of nodes respecting the given condition(s)}
\usage{
subset.counts(chia.subset, all.conditions, proportion = TRUE)
}
\arguments{
\item{chia.subset}{A list containing a graph and ChIA-PET regions, as returned by \code{\link{chia.vertex.subset}}.}

\item{proportion}{Should the number of occurences be converted to proportions?}

\item{conditions}{a string with the condition to respect}
}
\value{
A named vector (table) with the number of occurences (or proportions) of each category.
}
\description{
Counts the number of nodes respecting the given condition(s)
}

