% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{summarize_distances}
\alias{summarize_distances}
\title{Returns a summary of the distances between two sets of nodes.}
\usage{
summarize_distances(chia.obj, from, to, max = 3, values = FALSE)
}
\arguments{
\item{chia.obj}{The ChIA object on which distances must be computed.}

\item{from}{A vector of indices describing the starting nodes in the distance calculations.}

\item{to}{A vector of indices describing the destination nodes in the distance calculations.}

\item{max}{The maximum distance to be reported.}

\item{values}{Report absolute values instead of proportions.}
}
\value{
A vector describing the number of/proportion of nodes from 'from'
   with the corresponding distance to nodes in 'to'.
}
\description{
Returns a summary of the distances between two sets of nodes.
}

