% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricsChIA.R
\name{TF.presence}
\alias{TF.presence}
\title{Finds which transcription factors are present}
\usage{
TF.presence(chia.subset, number = FALSE)
}
\arguments{
\item{chia.subset}{A list containing a graph and ChIA-PET regions, as returned by \code{\link{chia.vertex.subset}}.}

\item{number}{Shoul the number of overlap be counted?}
}
\value{
A named vector with the presence or absence of each transcription factor.
}
\description{
Finds which transcription factors are present
}

