% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentChIA.R
\name{tf_enrichment}
\alias{tf_enrichment}
\title{Calculate enrichment of transcription factors on a selection of nodes.}
\usage{
tf_enrichment(chia.obj, draw.function)
}
\arguments{
\item{chia.obj}{The network on which TF enrichment needs to be performed.}

\item{draw.function}{A function which, given a ChIA object, returns a subset of the "drawn" nodes.}
}
\value{
A data-frame giving the enrichments of all transcription factors at the specified nodes.
}
\description{
Calculate enrichment of transcription factors on a selection of nodes.
}

