% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoherenceChIA.R
\name{expression.coherence}
\alias{expression.coherence}
\title{Generates a fucntion which takes a chia object, and determine which of its nodes
are in the top 25% and in the bottom 25% of expression.}
\usage{
expression.coherence(chia.obj, top, bottom)
}
\arguments{
\item{chia.obj}{The base ChIA object on which to assess gene activity.}

\item{top}{The expression value limit of the top 25% expressed genes.}

\item{bottom}{The expression value limit of the bottom 25% expressed genes.}
}
\value{
A 2-level vector with 
  "Top", "Bottom" or NA depending on the value of the given 
  fold-change column.
}
\description{
Generates a fucntion which takes a chia object, and determine which of its nodes
are in the top 25% and in the bottom 25% of expression.
}

