% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessorsChIA.R
\name{get.chips}
\alias{get.chips}
\title{Obtain a matrix of all possible ChIP data sets: transcription factors, polymerase-binding
and histone marks.}
\usage{
get.chips(chia.obj, overlap.threshold = 0)
}
\arguments{
\item{chia.obj}{A list containing the ChIA-PET data, as returned by \code{\link{load.chia}}.}
}
\value{
A matrix describing which epigenetic marks appears on which regions.
}
\description{
Combines the results of get.tf, get.polymerases and get.histones.
}

